fixed layanan and forum functions

+++ app/Http/Controllers/ForumController.php
@@ -11,6 +11,7 @@ use DateInterval;
 use App\Models\Forum;
 use App\Models\ForumVote;
 use App\Models\ForumMedia;
+use App\Models\ForumView;
 use Illuminate\Support\Facades\DB;
 
 class ForumController extends BaseController
@@ -22,14 +23,17 @@ class ForumController extends BaseController
     {
         $currentDate = date('Y-m-d H:i:s', strtotime(date('Y-m-1')));
         $date = new DateTime($currentDate);
-        $forumTerbaru = Forum::where('replied_to', null)->orderBy('created_at', 'desc')->get();
         if (Auth::user()->level == 2) {
+            $forumTerbaru = Forum::where('replied_to', null)->orderBy('created_at', 'desc')->get();
             $forumTrending = Forum::where('replied_to', null)->whereRaw('created_at between ? and ?', [$date->format('Y-m-d H:i:s'), $date->add(new DateInterval('P30D'))->format('Y-m-d H:i:s')])
                 ->orderBy('upvote_count', 'desc')->orderBy('view_count', 'desc')
                 ->get();
             return view('forum.index', ['forumTerbaru' => $forumTerbaru, 'forumTrending' => $forumTrending, 'page' => 'forum']);
         }
-        return view('admin.forum.index', ['forumTerbaru' => $forumTerbaru, 'page' => 'forum']);
+        $forumDone = Forum::where('status', 'selesai')->where('replied_to', null)->get();
+        $forumProcessed = Forum::where('status', 'proses')->where('replied_to', null)->get();
+        $forumUnprocessed = Forum::where('status', 'menunggu')->where('replied_to', null)->get();
+        return view('admin.forum.index', ['forumDone' => $forumDone, 'forumUnprocessed' => $forumUnprocessed, 'forumProcessed' => $forumUnprocessed, 'page' => 'forum']);
     }
 
     /**
@@ -47,25 +51,33 @@ class ForumController extends BaseController
     {
         $validator = Validator::make($request->all(), [
             'judul' => $id == null ? 'required|string' : 'nullable|string',
-            'deskripsi' => 'required|string',
+            'deskripsi' => Auth::user()->level == 1 ? 'nullable|string' : 'required|string',
             'gambar.*' => 'nullable|file|mimes:jpg,png,gif,jpeg|max:1500'
         ]);
         if ($validator->fails()) {
             return redirect('forum/' . ($id == null ? '' : $id))->with('error', $validator->errors()->first());
         }
 
+        $balasanTemplate = "keluh kesah anda telah kami baca dan akan kami tindak lanjuti, terimakasih banyak!";
+
         $create = Forum::create([
             'creator_id' => Auth::user()->id,
             'judul' => strlen($request->judul) == 0 ? "" : $request->judul,
-            'content' => $request->deskripsi,
-            'replied_to' => $id == null ? '' : $id,
+            'content' => strlen($request->deskripsi) != 0 ? $request->deskripsi : $balasanTemplate,
+            'replied_to' => $id == null ? null : $id,
         ]);
         $storeImage = $this->uploadFile($request->gambar, $create);
         if (!$storeImage) {
             $deleteImage = $this->deleteFile($id);
             return redirect('forum')->with('error', 'Gagal menghapus forum');
         }
-        return redirect('forum')->with('success', 'Berhasil membuat forum, silakan tunggu respon dari pihak desa');
+        if (Auth::user()->level == 1) {
+            Forum::where('id', $id)->update([
+                'status' => 'proses'
+            ]);
+            return redirect('admin/forums/' . $id)->with('success', 'Berhasil membalas forum');
+        }
+        return redirect($id == null ? 'forum' : 'forum/' . $id)->with('success', 'Berhasil membuat forum, silakan tunggu respon dari pihak desa');
     }
 
     /**
@@ -74,12 +86,21 @@ class ForumController extends BaseController
     public function show(string $id)
     {
         $forum = Forum::find($id);
-        if(Auth::user()->id != $forum->user->id){
+        $check = ForumView::where('forum_id', $id)->where('user_id', Auth::user()->id)->first();
+        if ($check == null) {
+            $create = ForumView::create([
+                'forum_id' => $id,
+                'user_id' => Auth::user()->id,
+            ]);
             $update = $forum->update([
                 'view_count' => $forum->view_count + 1
             ]);
         }
-        return view('forum.show', ['data' => Forum::find($id), 'replies' => Forum::where('replied_to', $id)->get(), 'page' => 'forum']);
+        if (Auth::user()->level == 1) {
+            return view('admin.forum.detail', ['data' => $forum, 'page' => 'forum']);
+        }
+        $replies = Forum::join('users', 'users.id', '=', 'forum.creator_id')->where('replied_to', $id)->orderBy('users.level', 'asc')->get();
+        return view('forum.show', ['data' => Forum::find($id), 'replies' => $replies, 'page' => 'forum']);
     }
 
     /**
@@ -124,10 +145,42 @@ class ForumController extends BaseController
     {
         $deleteImage = $this->deleteFile($id);
         if ($deleteImage) {
-            $deleteForum = Forum::where('id')->delete();
-            return redirect('forum')->with('success', 'Forum berhasil dihapus');
+            $deleteForum = Forum::destroy($id);
+            return redirect()->back()->with('success', 'Forum berhasil dihapus');
         }
-        return redirect('forum')->with('error', 'Forum gagal dihapus');
+        return redirect()->back()->with('error', 'Forum gagal dihapus');
+    }
+
+    public function changeStatus(Request $request, $id)
+    {
+        $update = Forum::where('id', $id)->update([
+            'is_ditutup' => 1,
+            'status' => $request->status
+        ]);
+        if ($update) {
+            return redirect('admin/forums/' . $id)->with('success', 'Berhasil mengubah status forum');
+        }
+        return redirect('admin/forums/' . $id)->with('error', 'gagal mengubah status forum');
+    }
+
+    public function changeForumOpen($id)
+    {
+        $find = Forum::find($id);
+        $message = '';
+        if ($find->is_ditutup == 0) {
+            $find->update([
+                'is_ditutup' => 1
+            ]);
+
+            $message = 'Forum berhasil ditutup';
+        } else {
+            $find->update([
+                'is_ditutup' => 0
+            ]);
+
+            $message = 'Forum berhasil dibuka';
+        }
+        return redirect('admin/forums/' . $id)->with('success', $message);
     }
 
     public function upvote($id)
@@ -148,12 +201,15 @@ class ForumController extends BaseController
                 'upvote_count' => $findForum->upvote_count + 1,
             ]);
         }
-        return redirect('forum/' . $id);
+        return redirect()->back();
     }
 
     public function uploadFile($items, Forum $forum)
     {
         $i = 1;
+        if ($items == null) {
+            return true;
+        }
         foreach ($items as $item) {
             $extension = $item->getClientOriginalExtension();
             $replacedTimestamps = str_replace(':', '-', $forum->created_at);
@@ -175,6 +231,9 @@ class ForumController extends BaseController
     public function deleteFile($id)
     {
         $filesPermohonan = ForumMedia::where('forum_id', $id)->get();
+        if (sizeof($filesPermohonan) == 0) {
+            return true;
+        }
         foreach ($filesPermohonan as $file) {
             $delete = Storage::disk('public')->delete('forum/' . $file->nama_file);
             if (!$delete) {
